fixed ResourceQuota values

+++ src/domain/dto/resource.quotas.dto.go
@@ -6,9 +6,9 @@ type ResourceQuotas struct {
 }
 
 type ResourceQuotasItem struct {
-	Name             string `json:"name"`
-	DisplayValue     string `json:"displayValue"`
-	DisplayUsedValue string `json:"displayUsedValue"`
-	LimitValue       int64  `json:"limitValue"`
-	UsedValue        int64  `json:"usedValue"`
+	Name             string  `json:"name"`
+	DisplayValue     string  `json:"displayValue"`
+	DisplayUsedValue string  `json:"displayUsedValue"`
+	LimitValue       float64 `json:"limitValue"`
+	UsedValue        float64 `json:"usedValue"`
 }
